edit account now survives configuration change

+++ app/src/main/java/com/example/fyp/menucreator/ui/fragments/SecondFragment.kt
@@ -1,11 +1,13 @@
 package com.example.fyp.menucreator.ui.fragments
 
 import android.app.AlertDialog
+import android.content.Context
 import android.os.Build
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.activity.OnBackPressedCallback
 import androidx.annotation.RequiresApi
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
@@ -62,10 +64,6 @@ class SecondFragment : Fragment() {
 
     private var action : NavDirections? = null
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-    }
-
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
@@ -95,15 +93,29 @@ class SecondFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
 
         binding.deleteFab.setOnClickListener{
-            deleteProduct()
+            deleteDialog()
         }
         binding.editFab.setOnClickListener{
             action?.let { it1 -> findNavController().navigate(it1) }
         }
     }
 
-    private fun navigateBack() = findNavController().navigate(SecondFragmentDirections.actionSecondFragmentToFirstFragment())
+    private fun deleteDialog(){
+        AlertDialog.Builder(context)
+            .setTitle("Delete Product")
+            .setMessage("Are you sure deleting this product?")
+            .setPositiveButton("Delete") { _, _ ->
+                deleteProduct()
+            }
+            .setNegativeButton("Cancel") { dialog, _ ->
+                dialog.dismiss()
+            }
+            .setCancelable(false)
+            .create()
+            .show()
+    }
 
+    private fun navigateBack() = findNavController().navigateUp()
     private fun deleteProduct() {
         observeFoodDeletion()
         viewModel.deleteProduct(productId!!)
